LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/aspect-ratio/replaced-elements-inside-grid-item-percent-height-and-padding.html

<!DOCTYPE html>
<html>
<head>
<link rel="author" title="Sammy Gill" href="mailto:sammy.gill@apple.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="help" href="https://drafts.csswg.org/css-sizing-4/#aspect-ratio">
<meta name="assert" content="Replaced elements' with no intrinsic size but aspect ratio width should remain same">
<style>
  grid {
    display: grid;
    width: 100px;
  }
  .aspect-ratio-and-padding {
    padding: 10px;
    aspect-ratio: 1
  }
  .percent-height {
    height: 100%;
  }
  .aspect-ratio {
    aspect-ratio: 1;
  }
</style>
</head>
<body>
  <grid>
    <div class="aspect-ratio-and-padding">
      <div class="percent-height">
        <svg class="percent-height" viewBox="0 0 1 1"></svg>
        <img class="percent-height aspect-ratio">
        <canvas class="percent-height aspect-ratio"></canvas>
      </div>
    </div>
  </grid>
</body>
<script>
document.body.offsetHeight;
let svg = document.querySelector("svg");
let initialSVGWidth = window.getComputedStyle(svg)["width"];

let canvas = document.querySelector("canvas");
let initialCanvasWidth = window.getComputedStyle(canvas)["width"];

let img = document.querySelector("img");
let initialImgWidth = window.getComputedStyle(img)["width"];

document.body.style.width = "90%";
document.body.offsetHeight;

test(() => {
  assert_equals(initialSVGWidth, window.getComputedStyle(svg)["width"]);
});

test(() => {
  assert_equals(initialImgWidth, window.getComputedStyle(img)["width"]);
});

test(() => {
  assert_equals(initialCanvasWidth, window.getComputedStyle(canvas)["width"]);
});
</script>
